requirements.txt
sqlalchemy~=2.0.34
bcrypt~=3.2.0
gradio~=4.44.0
fastapi~=0.112.2
starlette~=0.38.2
uvicorn~=0.20.0
langchain-core~=0.3.5
langchain-community~=0.3.0
python-dotenv~=1.0.1
pymysql
httpx_sse
pyjwt
neo4j~=5.25.0